git commit -m "day15 mid"

+++ day15/serverSide/controllers/feed.js
@@ -2,35 +2,29 @@ const fs = require("fs");
 const path = require("path");
 const {validationResult} = require("express-validator");
 const Post = require("../models/post");
+const User = require("../models/user");
 
-exports.getPosts = (req,res,next) => {
+exports.getPosts = async (req,res,next) => {
     const currentPage = req.query.page || 1;
     const perPage = 2;
     let totalItems;
-    Post.find()
-        .countDocuments()
-        .then(count => {
-            totalItems = count;
-            return Post
-                .find()
-                .skip((currentPage -1) * perPage)
-                .limit(perPage);
-        })
-        .then(posts => {
-            res
-                .status(200)
-                .json({
-                    message: "fetched posts",
-                    posts: posts,
-                    totalItems: totalItems
-                });
-        })
-        .catch(err => {
-            if(!err.statusCode){
-                err.statusCode = 500;
-            }
-            next(err);
+    try {
+        const totalItems = await Post.find().countDocuments();
+        const posts = await Post.find()
+            .skip((currentPage-1)*perPage)
+            .limit(perPage);
+        
+        res.status(200).json({
+            message: "fetched posts",
+            posts:posts,
+            totalItems:totalItems
         });
+    } catch(err) {
+        if(!err){
+            err.statusCode = 500
+        }
+        next(err);
+    }
 };
 
 exports.createPost = (req,res,next)=> {
@@ -48,18 +42,28 @@ exports.createPost = (req,res,next)=> {
     const imageUrl = req.file.path;
     const title = req.body.title;
     const content = req.body.content;
+    let creator;
     const post = new Post({
         title: title,
         content: content,
         imageUrl: imageUrl,
-        creator: {name: "feyz"} 
+        creator: req.userId
     });
     post
         .save()
+        .then(result=> {
+            return User.findById(req.userId);
+        })
+        .then(user=> {
+            creator = user;
+            user.posts.push(post);
+            return user.save();
+        })
         .then(result => {
             res.status(201).json({
                 message: "post created successfully",
-                post: result
+                post: post,
+                creator: { _id: creator._id, name: creator.name}
             });
         })
         .catch(err => {
@@ -81,7 +85,7 @@ exports.getPost = (req,res,next) => {
             }
             res.status(200).json({message: "post fetched", post: post });
         })
-        .catch(error => {
+        .catch(err => {
             if(!err.statusCode){
                 err.statusCode = 500;
             }
@@ -115,6 +119,11 @@ exports.updatePost = (req,res,next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            if(post.creator.toString() !== req.userId) {
+                const error = new Error("not auth");
+                error.statusCode = 403;
+                throw error;
+            }
             if(imageUrl !== post.imageUrl){
                 clearImage(post.imageUrl);
             }
@@ -143,8 +152,20 @@ exports.deltePost = (req,res,next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            if(post.creator.toString() !== req.userId) {
+                const error = new Error("not auth");
+                error.statusCode = 403;
+                throw error;
+            }
             clearImage(post.imageUrl);
             return Post.findByIdAndRemove(postId);
+        })
+        .then(result => {
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            user.posts.pull(postId);
+            return user.save();
         })
         .then(result => {
             res.status(200).json({message: "deleted"});
